requirements_versions.txt
transformers==4.19.2
diffusers==0.2.4
basicsr==1.3.5
gfpgan
gradio==3.3.1
numpy
Pillow==9.2.0
realesrgan==0.2.5.0
torch
transformers==4.19.2
omegaconf==2.1.1
pytorch_lightning==1.7.2
scikit-image==0.19.2
fonts
font-roboto
timm==0.4.12
fairscale
piexif==1.1.3

uvicorn==0.18.3
fastapi==0.85.0
pydantic==1.10.2